# fix(server): Check album asset membership in bulk