@adamantike 2023-10-23
Sure! The way I used to test these changes is to both add and remove photos from an album. I tested using **370 assets**, and created two small scripts:

* `script-add-and-delete.sh` added all assets to the album (initially empty), and then removed all of them in a subsequent request.

```sh
# Current (master):
$ hyperfine "./script-add-and-delete.sh"
Benchmark 1: ./script-add-and-delete.sh
  Time (mean ± σ):     660.4 ms ±  19.0 ms    [User: 3.8 ms, System: 5.4 ms]
  Range (min … max):   631.0 ms … 695.2 ms    10 runs

# With PR changes:
$ hyperfine "./script-add-and-delete.sh"
Benchmark 1: ./script-add-and-delete.sh
  Time (mean ± σ):     336.6 ms ±  21.6 ms    [User: 3.5 ms, System: 5.5 ms]
  Range (min … max):   321.9 ms … 393.0 ms    10 runs
```

* `script-delete-with-assets-not-present.sh` tried to remove all asset ids from the empty album. In this scenario, backend replies with `"success":false,"error":"not_found"` for all assets.

```sh
# Current (master):
$ hyperfine "./script-delete-with-assets-not-present.sh"
Benchmark 1: "./script-delete-with-assets-not-present.sh"
  Time (mean ± σ):     154.0 ms ±   4.8 ms    [User: 2.2 ms, System: 2.6 ms]
  Range (min … max):   147.4 ms … 169.1 ms    17 runs

# With PR changes:
$ hyperfine "./script-delete-with-assets-not-present.sh"
Benchmark 1: ./script-delete-with-assets-not-present.sh
  Time (mean ± σ):       8.7 ms ±   0.8 ms    [User: 1.8 ms, System: 2.0 ms]
  Range (min … max):     7.9 ms …  15.6 ms    178 runs

```